Extract file-related code to new FileManager class
