+++ src/modules/peek/Peek.UI/MainWindow.xaml
@@ -25,7 +25,7 @@
 
         <fp:FilePreview
             Grid.Row="1"
-            File="{x:Bind ViewModel.CurrentFile, Mode=OneWay}"
+            File="{x:Bind ViewModel.FileManager.CurrentFile, Mode=OneWay}"
             PreviewSizeChanged="FilePreviewer_PreviewSizeChanged" />
     </Grid>
 </winuiex:WindowEx>
